:sparkles: create useSlidingPuzzle composable
